requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
google-auth==2.28.1
requests>=2.31.0

# Settings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0

# OpenAI
openai==1.12.0

# HTTP client
httpx==0.26.0

# Development
watchfiles==0.21.0

# Video and Image Processing
moviepy>=1.0.3
Pillow>=10.0.0

